malcolm: pandas and pytorch loader
I've spearated the data loader in a abstract class, a pytorch loader,
and a pandas loader. this enable non pytorch people to easily load the
dataset and work on it.
