+++ .gitignore
@@ -1,0 +1,3 @@
+*.pyc
+*.kate-swp
+*.kdev4
